+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Lexing done by minimal lexer
+- Moved main.go to ./cmd/nofox/
 
 ## [0.0.1] - 2026-01-25